refactor: move ResponseRecorder from contracts to server
ResponseRecorder is an implementation, not a contract. BodyCapturer and
ResponseCapture (interface/type used by ResponseWriter) stay in contracts;
the concrete struct and constructor move to server where they belong.
