+++ internal/handler/har/middleware_test.go
@@ -18,6 +18,7 @@ import (
 
 	"github.com/evg4b/uncors/internal/contracts"
 	"github.com/evg4b/uncors/internal/handler/har"
+	"github.com/evg4b/uncors/internal/server"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -89,7 +90,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 		})
 
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, makeHARRequest(t, "http://example.com/path")) //nolint:errcheck
 
 		assert.True(t, called)
@@ -108,7 +109,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 		})
 
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, makeHARRequest(t, "http://example.com/")) //nolint:errcheck
 
 		require.NoError(t, harWriter.Close())
@@ -127,7 +128,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 
 		req := makeHARRequest(t, "http://example.com/search?q=foo&page=2")
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, req) //nolint:errcheck
 
 		require.NoError(t, harWriter.Close())
@@ -159,7 +160,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 		require.NoError(t, err)
 
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, req) //nolint:errcheck
 
 		assert.Equal(t, body, received)
@@ -181,7 +182,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 		req.Header.Set("Authorization", "Bearer eyJhbGciOiJSUzI1NiJ9")
 
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, req) //nolint:errcheck
 
 		require.NoError(t, harWriter.Close())
@@ -220,7 +221,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 		req.TLS = &tls.ConnectionState{}
 
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, req) //nolint:errcheck
 
 		require.NoError(t, harWriter.Close())
@@ -249,7 +250,7 @@ func TestMiddleware_Wrap(t *testing.T) {
 		req.Header.Set("Authorization", "Bearer token123")
 
 		rec := httptest.NewRecorder()
-		rr := contracts.NewResponseRecorder(rec)
+		rr := server.NewResponseRecorder(rec)
 		mdlw.Wrap(next).ServeHTTP(rr, req) //nolint:errcheck
 
 		require.NoError(t, harWriter.Close())
@@ -302,7 +303,7 @@ func TestMiddleware_Wrap_Decompression(t *testing.T) {
 			})
 
 			rec := httptest.NewRecorder()
-			rr := contracts.NewResponseRecorder(rec)
+			rr := server.NewResponseRecorder(rec)
 			mdlw.Wrap(next).ServeHTTP(rr, makeHARRequest(t, "http://example.com/api")) //nolint:errcheck
 
 			require.NoError(t, harWriter.Close())
