+++ internal/server/recorder_test.go
@@ -1,4 +1,4 @@
-package contracts_test
+package server_test
 
 import (
 	"net/http"
@@ -7,19 +7,20 @@ import (
 	"time"
 
 	"github.com/evg4b/uncors/internal/contracts"
+	"github.com/evg4b/uncors/internal/server"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestResponseRecorder_StatusCode(t *testing.T) {
 	t.Run("returns 0 by default", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 
 		assert.Equal(t, 0, rec.StatusCode())
 	})
 
 	t.Run("returns code set by WriteHeader", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 		rec.WriteHeader(http.StatusNotFound)
 
 		assert.Equal(t, http.StatusNotFound, rec.StatusCode())
@@ -29,7 +30,7 @@ func TestResponseRecorder_StatusCode(t *testing.T) {
 func TestResponseRecorder_Write(t *testing.T) {
 	t.Run("writes through to underlying writer", func(t *testing.T) {
 		underlying := httptest.NewRecorder()
-		rec := contracts.NewResponseRecorder(underlying)
+		rec := server.NewResponseRecorder(underlying)
 
 		_, err := rec.Write([]byte("hello"))
 		require.NoError(t, err)
@@ -39,7 +40,7 @@ func TestResponseRecorder_Write(t *testing.T) {
 
 	t.Run("buffers body and still writes through when capture is enabled", func(t *testing.T) {
 		underlying := httptest.NewRecorder()
-		rec := contracts.NewResponseRecorder(underlying)
+		rec := server.NewResponseRecorder(underlying)
 		rec.EnableBodyCapture()
 
 		_, err := rec.Write([]byte("buffered"))
@@ -51,27 +52,27 @@ func TestResponseRecorder_Write(t *testing.T) {
 
 func TestResponseRecorder_Captured(t *testing.T) {
 	t.Run("returns correct status code", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 		rec.WriteHeader(http.StatusCreated)
 
 		assert.Equal(t, http.StatusCreated, rec.Captured().StatusCode)
 	})
 
 	t.Run("normalises missing WriteHeader to 200", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 
 		assert.Equal(t, http.StatusOK, rec.Captured().StatusCode)
 	})
 
 	t.Run("body is nil when capture not enabled", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 		_, _ = rec.Write([]byte("ignored"))
 
 		assert.Nil(t, rec.Captured().Body)
 	})
 
 	t.Run("body is captured when EnableBodyCapture is called", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 		rec.EnableBodyCapture()
 		_, _ = rec.Write([]byte("captured"))
 
@@ -79,7 +80,7 @@ func TestResponseRecorder_Captured(t *testing.T) {
 	})
 
 	t.Run("second EnableBodyCapture call is a no-op", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 		rec.EnableBodyCapture()
 		rec.EnableBodyCapture()
 		_, _ = rec.Write([]byte("once"))
@@ -88,7 +89,7 @@ func TestResponseRecorder_Captured(t *testing.T) {
 	})
 
 	t.Run("duration is non-zero", func(t *testing.T) {
-		rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+		rec := server.NewResponseRecorder(httptest.NewRecorder())
 
 		time.Sleep(time.Millisecond)
 
@@ -97,7 +98,7 @@ func TestResponseRecorder_Captured(t *testing.T) {
 }
 
 func TestResponseRecorder_ImplementsInterfaces(t *testing.T) {
-	rec := contracts.NewResponseRecorder(httptest.NewRecorder())
+	rec := server.NewResponseRecorder(httptest.NewRecorder())
 
 	t.Run("implements ResponseWriter", func(_ *testing.T) {
 		var _ contracts.ResponseWriter = rec
